Modified tiles to only allow paths on tiles of height 2 and 3
### Disclaimer
Path generation now decides height of underlying nodes. 
This is contrary to how we planned initially, as we thought about generating the height first and then having the path try to generate around the landmarks. However, this would be a lot more implementation work and would also require more regenerations before a map with a solvable path is generated.
The way it is implemented now, the success rate of generating a path is way higher, but the path will be less interesting.

### Description
- Removed the tiles Height1_Path and Height2_Path from all valid neighbour tiles lists
- Removed the tiles Height1_Path and Height2_Path from allTiles, thereby removing them from all potentialTiles, preventing them from spawning entirely